sending through socket

+++ ost/model/ost-header.h
@@ -7,35 +7,36 @@
 #include "ns3/header.h"
 #include "spw_packet.h"
 
-namespace ns3 {
+namespace ns3
+{
 
-typedef enum {
-    ACK,
-    SYN,
-    RST,
-    DTA // virtual
-} SegmentFlag;
+    typedef enum
+    {
+        ACK,
+        SYN,
+        RST,
+        DTA // virtual
+    } SegmentFlag;
 
-class OstHeader: public Header {
-    typedef uint8_t FlagOctet;
+    class OstHeader : public Header
+    {
+        typedef uint8_t FlagOctet;
 
     public:
         OstHeader()
             : flags(0),
-            seq_number(0),
-            source_addr(0),
-            payload_length(0)
-            {};
-        OstHeader(uint8_t seq_n, uint8_t addr, uint8_t len)
+              seq_number(0),
+              source_addr(0),
+              payload_length(0){};
+        OstHeader(uint8_t seq_n, uint8_t addr, uint16_t len)
             : seq_number(seq_n),
-            source_addr(addr),
-            payload_length(len)
-            {};
-        ~OstHeader() {};
+              source_addr(addr),
+              payload_length(len){};
+        ~OstHeader(){};
 
         static TypeId GetTypeId();
         TypeId GetInstanceTypeId() const override;
-        void Print(std::ostream& os) const override;
+        void Print(std::ostream &os) const override;
         uint32_t GetSerializedSize() const override;
         void Serialize(Buffer::Iterator start) const override;
         uint32_t Deserialize(Buffer::Iterator start) override;
@@ -46,8 +47,8 @@ class OstHeader: public Header {
         void set_src_addr(uint8_t);
         uint8_t get_src_addr();
 
-        void set_payload_len(uint8_t);
-        uint8_t get_payload_len();
+        void set_payload_len(uint16_t);
+        uint16_t get_payload_len();
 
         void set_flag(SegmentFlag flag);
         void unset_flat(SegmentFlag flag);
@@ -56,12 +57,13 @@ class OstHeader: public Header {
         bool is_syn();
         bool is_rst();
         bool is_dta();
+
     private:
         FlagOctet flags;
         uint8_t seq_number;
         uint8_t source_addr;
         uint16_t payload_length;
-};
+    };
 
 }
 #endif
